Add shareable shopping list for meal planning
